requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.10.0
attrs==25.4.0
cffi==2.0.0
Django==4.2.8
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pycparser==2.23
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
